Adding type annotation to map visitor in desugar and consolidated string dict to map functions
