+++ src/arr/compiler/ts-codegen-helpers.ts
@@ -1,5 +1,5 @@
 import * as J from 'estree';
-import type { List } from './ts-impl-types';
+import type { List, StringDict, MutableStringDict, PFunction, PMethod } from './ts-impl-types';
 import type * as A from './ts-ast';
 
 export type Variant<T, V> = T & { $name: V };
@@ -7,6 +7,23 @@ export type Variant<T, V> = T & { $name: V };
 export type PyretObject = {
   dict: any
 };
+
+type SDExports = {
+  dict: { values: { dict: {
+    'make-mutable-string-dict': PFunction<<T>() => MutableStringDict<T>>
+    'is-mutable-string-dict': PFunction<(val: any) => boolean>,
+    'make-string-dict': PFunction<<T>() => StringDict<T>>,
+    'is-string-dict': PFunction<(val: any) => boolean>,
+    'map-keys': PFunction<<T, U>(f: ((key: T) => U), isd: StringDict<T>) => List<U>>,
+    'map-keys-now': PFunction<<T, U>(f: ((key: T) => U), msd: MutableStringDict<T>) => List<U>>,
+    'fold-keys': PFunction<<T, U>(f: (key: string, acc: U) => U, init: U, isd: StringDict<T>) => U>,
+    'fold-keys-now': PFunction<<T, U>(f: (key: string, acc: U) => U, init: U, msd: MutableStringDict<T>) => U>,
+    'each-key': PFunction<<T>(f: ((key: T) => void), isd: StringDict<T>) => void>,
+    'each-key-now': PFunction<<T>(f: ((key: T) => void), msd: MutableStringDict<T>) => void>,
+  }}}
+}
+
+type DropFirst<T extends unknown[]> = ((...p: T) => void) extends ((p1: infer P1, ...rest: infer R) => void) ? R : never
 
 /** The supertype of all possible Pyret data values. */
 // TODO: add the remaining reflective accessors to this definition
@@ -82,6 +99,11 @@ export interface Exports {
       v : Visitor<T>,
       d : T,
     ) => Ret,
+    mapFromStringDict: <T>(s : StringDict<T>) => Map<string, T>,
+    mapFromMutableStringDict: <T>(s : MutableStringDict<T>) => Map<string, T>,
+    stringDictFromMap: <T>(m : Map<string, T>) => StringDict<T>,
+    mutableStringDictFromMap: <T>(m : Map<string, T>) => MutableStringDict<T>,
+    callMethod: <Name extends string, O extends {dict: {[n in Name]: PMethod<any, (...args: any[]) => any>}}>(obj : O, name: Name, ...args: DropFirst<Parameters<O["dict"][Name]["full_meth"]>>) => ReturnType<O["dict"][Name]["full_meth"]>,
 }
 
 ({
@@ -131,7 +153,7 @@ export interface Exports {
       compileSrcloc: 'tany',
     },
   },
-  theModule: function(runtime, _, __, Ain: A.Exports) {
+  theModule: function(runtime, _, __, Ain: A.Exports, SD: SDExports) {
     const A = Ain.dict.values.dict;
     class ExhaustiveSwitchError extends Error {
       constructor(v: never, message?: string) {
@@ -154,40 +176,6 @@ export interface Exports {
       }
     }
 
-    function visit<T extends { $name: string, dict: {} }>(v : Partial<Record<T["$name"], any>>, d : T) {
-      if(typeof d !== "object" || !("$name" in d)) { throw new Error("Visit failed: " + JSON.stringify(d)); }
-      if(d.$name in v) { v[d.$name](v, d); }
-      else {
-        for(const [k, subd] of Object.entries(d.dict)) {
-          if(typeof subd === 'object' && "$name" in subd) {
-            visit(v, subd as any);
-          }
-        }
-      }
-    }
-
-    function map<T extends { $name: string, dict: {} }, A extends T>(v : Partial<Record<T["$name"], any>>, d : A) : A {
-      if(typeof d !== "object" || !("$name" in d)) { throw new Error("Map failed: " + JSON.stringify(d)); }
-      if(d.$name in v) { return v[d.$name](v, d); }
-      else {
-        const newObj : typeof d = Object.create(Object.getPrototypeOf(d));
-        for(const [k, meta] of Object.entries(d)) {
-          if(k !== "dict") { newObj[k] = meta; }
-        }
-        newObj.dict = Object.create(Object.getPrototypeOf(d.dict));
-        for(const [k, subd] of Object.entries(d.dict)) {
-          if(typeof subd === 'object' && "$name" in subd) {
-            const result = map(v, subd as any);
-            newObj.dict[k] = result;
-          }
-          else {
-            newObj.dict[k] = subd;
-          }
-        }
-        return newObj;
-      }
-    }
-    
     function ExpressionStatement(e : J.Expression) : J.ExpressionStatement {
       return { type: "ExpressionStatement", expression: e };
     }
@@ -572,7 +560,33 @@ export interface Exports {
       }
     }
 
-
+    function callMethod<Name extends string, O extends {dict: {[n in Name]: PMethod<any, (...args: any[]) => any>}}>(obj : O, name: Name, ...args: DropFirst<Parameters<O["dict"][Name]["full_meth"]>>) : ReturnType<O["dict"][Name]["full_meth"]> {
+      return obj.dict[name].full_meth(obj, ...args);
+    }
+    function mapFromStringDict<T>(s : StringDict<T>) : Map<string, T> {
+      const m : Map<string, T> = new Map();
+      for (let valKey of listToArray(callMethod(s, 'keys-list'))) {
+        m.set(valKey, callMethod(s, "get-value", valKey));
+      }
+      return m;
+    }
+    function mapFromMutableStringDict<T>(s : MutableStringDict<T>) : Map<string, T> {
+      const m : Map<string, T> = new Map();
+      for (let valKey of listToArray(callMethod(s, 'keys-list-now'))) {
+        m.set(valKey, callMethod(s, "get-value-now", valKey));
+      }
+      return m;
+    }
+    function stringDictFromMap<T>(m : Map<string, T>): StringDict<T> {
+      return callMethod(mutableStringDictFromMap(m), 'freeze');
+    }
+    function mutableStringDictFromMap<T>(m : Map<string, T>): MutableStringDict<T> {
+      const s = SD.dict.values.dict['make-mutable-string-dict'].app<T>();
+      for (const [k, v] of m.entries()) {
+        callMethod(s, 'set-now', k, v);
+      }
+      return s;
+    }
 
     return runtime.makeJSModuleReturn({
       ArrayExpression,
@@ -617,6 +631,11 @@ export interface Exports {
       formatSrcloc,
       visit,
       map,
+      mapFromStringDict,
+      mapFromMutableStringDict,
+      stringDictFromMap,
+      mutableStringDictFromMap,
+      callMethod,
     });
   }
 })
