+++ src/arr/compiler/ts-direct-codegen.ts
@@ -63,7 +63,6 @@ import type { Variant, PyretObject } from './ts-codegen-helpers';
       compileSrcloc,
       dummyLoc,
       listToArray,
-      map,
       nameToKey,
       nameToName,
       nameToSourceString,
